app/backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
fastmcp
pydantic
websockets
aiohttp